move TestMain to main_test.go
